// apiConnection/gen/models/query-get.ts
/* tslint:disable */
/* eslint-disable */
/**
 * LCT Hack Backend
 * Backend part for LCT Hack
 *
 * OpenAPI spec version: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { SubQueryGet } from './sub-query-get';
/**
 * 
 * @export
 * @interface QueryGet
 */
export interface QueryGet {
    /**
     * Название запроса
     * @type {string}
     * @memberof QueryGet
     */
    name?: string;
    /**
     * Ссылка на файл с входными данными
     * @type {string}
     * @memberof QueryGet
     */
    inputFile: string;
    /**
     * Ссылка на файл с выходными данными
     * @type {string}
     * @memberof QueryGet
     */
    outputFile?: string;
    /**
     * Уникальный идентификатор записи
     * @type {string}
     * @memberof QueryGet
     */
    guid: string;
    /**
     * Список подзапросов
     * @type {Array<SubQueryGet>}
     * @memberof QueryGet
     */
    subQueries: Array<SubQueryGet>;
    /**
     * Уникальный идентификатор пользователя, создавшего запись
     * @type {string}
     * @memberof QueryGet
     */
    createdBy: string;
    /**
     * Уникальный идентификатор пользователя, обновившего запись
     * @type {string}
     * @memberof QueryGet
     */
    updatedBy: string;
    /**
     * Время создания записи
     * @type {Date}
     * @memberof QueryGet
     */
    createdAt: Date;
    /**
     * Время последнего обновления записи
     * @type {Date}
     * @memberof QueryGet
     */
    updatedAt: Date;
}